+++ src/main/java/com/github/minecraftschurlimods/bibliocraft/content/bigbook/BigBookItem.java
@@ -21,11 +21,11 @@ public class BigBookItem extends Item {
     public BigBookItem(boolean isWritten) {
         super(isWritten
                 ? new Properties()
-                .component(DataComponents.MAX_STACK_SIZE, 1)
+                .stacksTo(1)
                 .component(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, true)
                 .component(BCDataComponents.WRITTEN_BIG_BOOK_CONTENT, WrittenBigBookContent.DEFAULT)
                 : new Properties()
-                .component(DataComponents.MAX_STACK_SIZE, 1)
+                .stacksTo(1)
                 .component(BCDataComponents.BIG_BOOK_CONTENT, BigBookContent.DEFAULT)
         );
     }